Updated graphing and fixed index bug
