+++ AI_Tower_Defense/src/constants/aiConstants.py
@@ -14,7 +14,7 @@ EPSILON        = 0.1
 EPSILON_STEP   = 0.004
 EPSILON_PERIOD = 100
 LEARN_RATE            = 0.1
-N_EPISODES     = 2000
+N_EPISODES     = 10000
 M_STEPS        = 500
 
 # GA constants
